[feat] add uva 12503 12015 11679

+++ Sources/hxy/src/Volume101/p10114.c
@@ -81,7 +81,7 @@
 #include "Volume101/p10114.h"
 
 struct p10114_Depreciation {
-  Int month;
+  Int64 month;
   Double rate;
 };
